# MTLS_UI/src/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/quizzes": {
            "post": {
                "operationId": "QuizController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateQuizDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bài kiểm tra đã được tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "summary": "Tạo bài kiểm tra mới",
                "tags": ["quizzes"]
            },
            "get": {
                "operationId": "QuizController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách quiz"
                    }
                },
                "summary": "Lấy danh sách tất cả quiz",
                "tags": ["quizzes"]
            }
        },
        "/quizzes/{id}": {
            "get": {
                "operationId": "QuizController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài kiểm tra",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thông tin bài kiểm tra"
                    },
                    "404": {
                        "description": "Không tìm thấy bài kiểm tra"
                    }
                },
                "summary": "Lấy thông tin bài kiểm tra theo ID",
                "tags": ["quizzes"]
            },
            "put": {
                "operationId": "QuizController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài kiểm tra",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateQuizDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bài kiểm tra đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy bài kiểm tra"
                    }
                },
                "summary": "Cập nhật bài kiểm tra",
                "tags": ["quizzes"]
            },
            "delete": {
                "operationId": "QuizController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài kiểm tra",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Bài kiểm tra đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy bài kiểm tra"
                    }
                },
                "summary": "Xóa bài kiểm tra",
                "tags": ["quizzes"]
            }
        },
        "/quizzes/section/{id}": {
            "get": {
                "operationId": "QuizController_findBySectionId",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của section",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách bài kiểm tra"
                    },
                    "404": {
                        "description": "Không tìm thấy bài kiểm tra"
                    }
                },
                "summary": "Lấy danh sách bài kiểm tra theo ID của section",
                "tags": ["quizzes"]
            }
        },
        "/questions": {
            "post": {
                "operationId": "QuestionController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateQuestionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Câu hỏi đã được tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "summary": "Tạo câu hỏi mới",
                "tags": ["questions"]
            },
            "get": {
                "operationId": "QuestionController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách câu hỏi"
                    }
                },
                "summary": "Lấy danh sách tất cả câu hỏi",
                "tags": ["questions"]
            }
        },
        "/questions/{id}": {
            "get": {
                "operationId": "QuestionController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của câu hỏi",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thông tin câu hỏi"
                    },
                    "404": {
                        "description": "Không tìm thấy câu hỏi"
                    }
                },
                "summary": "Lấy thông tin câu hỏi theo ID",
                "tags": ["questions"]
            },
            "put": {
                "operationId": "QuestionController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của câu hỏi",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateQuestionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Câu hỏi đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy câu hỏi"
                    }
                },
                "summary": "Cập nhật câu hỏi",
                "tags": ["questions"]
            },
            "delete": {
                "operationId": "QuestionController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của câu hỏi",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Câu hỏi đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy câu hỏi"
                    }
                },
                "summary": "Xóa câu hỏi",
                "tags": ["questions"]
            }
        },
        "/questions/quiz/{id}": {
            "get": {
                "operationId": "QuestionController_findByQuizId",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của quiz",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách câu hỏi"
                    },
                    "404": {
                        "description": "Không tìm thấy câu hỏi"
                    }
                },
                "summary": "Lấy danh sách câu hỏi theo ID của quiz",
                "tags": ["questions"]
            }
        },
        "/questions/exercise/{id}": {
            "get": {
                "operationId": "QuestionController_findByExerciseId",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của exercise",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách câu hỏi"
                    },
                    "404": {
                        "description": "Không tìm thấy câu hỏi"
                    }
                },
                "summary": "Lấy danh sách câu hỏi theo ID của exercise",
                "tags": ["questions"]
            }
        },
        "/questions/batch": {
            "post": {
                "description": "Tạo nhiều câu hỏi thuộc quiz hoặc exercise cùng lúc",
                "operationId": "QuestionController_createMultiple",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMultipleQuestionsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Các câu hỏi đã được tạo thành công"
                    }
                },
                "summary": "Tạo nhiều câu hỏi cùng lúc",
                "tags": ["questions"]
            }
        },
...
        "/options/question/{id}": {
            "get": {
                "operationId": "OptionController_findListOptionByQuestionId",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của câu hỏi",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách tùy chọn"
                    },
                    "404": {
                        "description": "Không tìm thấy tùy chọn"
                    }
                },
                "summary": "Lấy danh sách tùy chọn theo ID của câu hỏi",
                "tags": ["options"]
            }
        },
...
        "/do-quizzes/quiz/{quizId}": {
            "get": {
                "operationId": "DoQuizController_findByQuizId",
                "parameters": [{
                    "name": "quizId",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài kiểm tra",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách bài làm của bài kiểm tra"
                    }
                },
                "summary": "Lấy danh sách bài làm quiz theo quiz ID",
                "tags": ["do-quizzes"]
            }
        },
...
        "/do-quizzes/student/{studentId}/quiz/{quizId}": {
            "get": {
                "operationId": "DoQuizController_findByStudentIdAndQuizId",
                "parameters": [{
                        "name": "studentId",
                        "required": true,
                        "in": "path",
                        "description": "ID của học sinh",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quizId",
                        "required": true,
                        "in": "path",
                        "description": "ID của bài kiểm tra",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về bài làm quiz của học sinh"
                    }
                },
                "summary": "Lấy bài làm quiz của học sinh theo quiz ID",
                "tags": ["do-quizzes"]
            }
        },
...
    },
    "info": {
        "title": "API Documentation",
        "description": "The API's MTLS description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateQuizDto": {
                "type": "object",
                "properties": {
                    "section": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "title": {
                        "type": "string",
                        "example": "Music Theory Basics"
                    },
                    "description": {
                        "type": "string",
                        "example": "A quiz to test your basic knowledge of music theory."
                    },
                    "duration": {
                        "type": "number",
                        "example": 30
                    }
                },
                "required": ["section", "title", "description", "duration"]
            },
            "UpdateQuizDto": {
                "type": "object",
                "properties": {
                    "section": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "title": {
                        "type": "string",
                        "example": "Advanced Music Theory"
                    },
                    "description": {
                        "type": "string",
                        "example": "An advanced quiz to test your knowledge of music theory."
                    },
                    "duration": {
                        "type": "number",
                        "example": 45
                    },
                    "questions": {
                        "example": ["60d21b4667d0d8992e610c85", "60d21b4667d0d8992e610c86"],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "CreateQuestionDto": {
                "type": "object",
                "properties": {
                    "quiz": {
                        "type": "string",
                        "description": "ID của quiz",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "exercise": {
                        "type": "string",
                        "description": "ID của bài tập nếu câu hỏi thuộc bài tập",
                        "example": "60d21b4667d0d8992e610c88"
                    },
                    "questionText": {
                        "type": "string",
                        "description": "Nội dung câu hỏi",
                        "example": "What is the capital of France?"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Trạng thái hoạt động của câu hỏi",
                        "example": true,
                        "default": true
                    },
                    "notation": {
                        "type": "string",
                        "description": "Chuỗi notation để hiển thị khuông nhạc",
                        "example": "c/4/q, e/4/q, g/4/q"
                    },
                    "questionType": {
                        "type": "string",
                        "description": "Loại câu hỏi",
                        "example": "multipleChoice",
                        "default": "multipleChoice",
                        "enum": ["multipleChoice", "vexFlow"]
                    }
                },
                "required": ["quiz", "questionText"]
            },
            "UpdateQuestionDto": {
                "type": "object",
                "properties": {
                    "quiz": {
                        "type": "string",
                        "description": "ID của quiz",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "exercise": {
                        "type": "string",
                        "description": "ID của bài tập nếu câu hỏi thuộc bài tập",
                        "example": "60d21b4667d0d8992e610c88"
                    },
                    "questionText": {
                        "type": "string",
                        "description": "Nội dung câu hỏi",
                        "example": "What is the capital of France?"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Trạng thái hoạt động của câu hỏi",
                        "example": true,
                        "default": true
                    },
                    "notation": {
                        "type": "string",
                        "description": "Chuỗi notation để hiển thị khuông nhạc",
                        "example": "c/4/q, e/4/q, g/4/q"
                    },
                    "questionType": {
                        "type": "string",
                        "description": "Loại câu hỏi",
                        "example": "multipleChoice",
                        "default": "multipleChoice",
                        "enum": ["multipleChoice", "vexFlow"]
                    }
                }
            },
            "CreateMultipleQuestionsDto": {
                "type": "object",
                "properties": {
                    "questions": {
                        "description": "Danh sách các câu hỏi cần tạo",
                        "example": [{
                                "quiz": "60d21b4667d0d8992e610c85",
                                "questionText": "What is the relative minor of C major?",
                                "notation": "c/4/q, e/4/q, g/4/q"
                            },
                            {
                                "exercise": "60d21b4667d0d8992e610c90",
                                "questionText": "What is the dominant of G major?",
                                "notation": "d/4/q, f#/4/q, a/4/q"
                            }
                        ],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateQuestionDto"
                        }
                    }
                },
                "required": ["questions"]
            },
...
        }
    }
}